requirements.txt
asgiref==3.7.2
boto3==1.34.21
botocore==1.34.21
certifi==2023.11.17
channels==4.2.0
charset-normalizer==3.3.2
Django==5.0.1
django-channels==0.7.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-model-utils==5.0.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.0.7
xmltodict==0.13.0